Use long long instead of LONGLONG - these may differ

+++ astrotcl/generic/Fits_IO.h
@@ -168,7 +168,7 @@ public:
     int get(const char* keyword, float& val) const;
     int get(const char* keyword, int& val) const;
     int get(const char* keyword, long& val) const;
-    int get(const char* keyword, LONGLONG& val) const;
+    int get(const char* keyword, long long& val) const;
     int get(const char* keyword, unsigned char& val) const;
     int get(const char* keyword, unsigned short& val) const;
     int get(const char* keyword, short& val) const;
